Fix use of calendarIds when creating events
This fixes how calendarIds are used when creating new events by using
the changes done in icalendar_vcard. The calendar should now be scanned
for correctly and if not found, a default calendar should be used.

Moreover, the data accessor now returns a string conaining the calendarId
and object uri after an event has been added successfully.